Camaro ECM system.  Beginning only with the temperature sensors.
This includes an early version of Synapse316 MCU which is usable.  It writes a character repeatedly to a UART TX that's in a different clock domain.  316 dispenses with muxb, so it has only one data muxer (data bus).  It's far easier to program and easier to expand to large i/o space than the previous 216 model.  And it does about 2x the work of a Zen216 in a given time frame.

This version contains a vestigial Qsys system that doesn't completely work.  e.g. not sure it can write to a real UART OK.

git-svn-id: svn://Laptop:port/personal/2014/camaro_ecm@389 58d5e4a1-9bfd-334e-a56e-dcfcd10d7f08