Speed tests on a Cyclone 3.
showed the I/O expander is only about 1% faster that the equivalent MCU register space, when that is of maximum size 48.
mockup peripherals were used, with realistic register allocation between the two.
so no need to use an expander, nor to add debugger support for it at this time.
the expander does allow a far larger 64k address space vs. 48 words in the MCU.
but the MCU and debugger already support Avalon for that.